backend/requirements.txt
fastapi
uvicorn
python-socketio[asyncio]
redis
asyncio
aioredis